deign
Each text of any substantial length, regardless of genre, invariably contains a genomic sequence. From the Holy Bible to the Egyptian Book of the Dead, to the corpus of text messages we send back and forth every day, to the very words we utter over the course of a lifetime, we, each of us, speaks life.

This trifle, deign, itemizes the abbreviations for the four nucleotides of deoxyribonucleic acid (DNA). Those four chemicals, Thymine, Guanine, Adenine and Cytosine, are abbreviated by the letters, T, G, A and C. It may also be augmented to produce a messenger RNA (mRNA) sequence by replacing the T for Thymine with U for Uracil.

deign extracts the four letters from a body of text, effectually transubstantiating the word from word to flesh.

The sequence of nucleotides, once extracted from a text, can be entered into a genome database such as the National Library of Medicine's Basic Local Alignment Search Tool (BLAST), the DNA Database of Japan (DDBJ), or the European Bioinformatics Institute (EBI), to name a few.

These databases compare the genetic sequence queries entered against a litany of genetic sequences gleaned from flora and fauna found on our planet. deign makes it possible to compare and examine the DNA or RNA sequences a text shares with a living organism. And, properly altered, one may also view the building-block aminos contained within a text by grouping the DNA or RNA codons by threes and converting them to their corresponding proteins with a dictionary-key mechanism. I will upload this dynamo soon.

So what? What is the point of this, you may ask? Well, I don't have a straightforward answer for that yet. In essence, I simply don't know enough about the scientific and medical worlds to know if this application is worth it's weight and wait (I've been sitting on this idea for damn near nine years.

I do have one anecdotal example in which I transubstantiated the Papyrus of Ani and found edible organisms of such significant but cohesive variety that I was able to create a nice recipe, a ceviche, to be exact.

Lofty though my intentions may be, I believe we can learn more than what is offered at face value in our shared cultural affinity for text. This union of Science, Technology, Engineering and logical Mathematics (STEM) with the Liberal and Language Arts, I believe, bears potential. How much is up to you, me and anyone who deigns to get down to the basics.

This tableau reliance restores a nation's laughter—edifies, teeming, a lingual legend.